Replace as many occurrences as possible of master by main
This is a consequence of the renaming of the master branch to main but I
also checked all the external URLs to see if they were changed or not. I
updated them if it was the case.

There are still a some problematic occurrences but they are mostly
related to external projects.